backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
gunicorn
mongoengine
dnspython
pymongo
python-dotenv
whitenoise
